move test into existing test_cli_agent.py
